Create README.md
Hello everyone. I am continuing to develop my game, Blitz Survivor: Tank Battle. I was planning to add a new ability to my game. This ability would create an object rotating around the tank object and decrease the health of enemy objects it collides with. To control the rotations, I did a brief research and decided to use the cosine and sine values available in Unity's Mathf class. To materialize this ability, I created simulations of a clock and a planet. I adjusted the hour, minute, and second hands proportionally in the simulation.